// src/interfaces/cart.interface.ts
export interface CartResponse {
    status: boolean;
    message: string;
    numOfCartItems: number;
    cartId: string;
    data: Cart;
}

export interface Cart {
    _id: string;
    cartOwner: string;
    products: CartItem[];
    createdAt: string;
    updatedAt: string;
    __v: number;
    totalCartPrice: number;
}

export interface CartItem {
    _id: string;
    count: number;
    price: number;
    product: Product;
}

export interface Product {
    _id: string;
    id: string;
    title: string;
    slug: string;
    quantity: number;
    imageCover: string;
    category: Category;
    brand: Brand;
    subcategory: SubCategory[];
    ratingsAverage: number;
}

export interface Category {
    _id: string;
    name: string;
    slug: string;
    image: string;
}

export interface Brand {
    _id: string;
    name: string;
    slug: string;
    image: string;
}

export interface SubCategory {
    _id: string;
    name: string;
    slug: string;
    category: string;
}